Add Finger-Cox-Jephcoat asymmetry using pseudo-Voight peak
Add new peak function based on that described in  Finger, Cox and Jephcoat (1994)
 J. Appl. Cryst. vol 27, pp 892-900.  An improved form of the normalisation
derivative (D in the above paper) as described in Hester (2013), J. Appl. Cryst.
(submitted) is included together with the accurate derivatives for H
and S described in the latter paper.  This asymmetry is essential for
modelling peaks from powder neutron diffraction facilities, and will also
improve fits to synchrotron and lab powder data.

In order to create the convoluted peakshape, Gauss-Legendre quadrature points
and weights are also included in a separate include file to avoid cluttering
the main code.

Signed-off-by: Marcin Wojdyr <wojdyr@gmail.com>