fix GUI problem with background stored in saved session, new cmd 'ui'
problem was:
Loading of .fit session files - I found that if you load a .fit session
file with a background function that has been subtracted, the software
doesnt know to have the bg subtract button depressed. So it doesnt allow
you to modify the bg once you load the session back in.

the new 'ui' command can be used to store status of the GUI. It
is experimental and not documented yet.