model.rst: remove confusion of F as sum vs F as set/list
It's a kind of confusing to overload F to stand for both set f_i and sum of them, as implied by formula https://fityk.nieto.pl/model.html#model-formula
Since the syntax ` F = %f +%g`, and output of `info F`, already implies F is sum of functions,
I think it's better to choose F to mean just sum of f_i, similarly for Z, and thus omit in docs the references to F/Z as sets/lists. (even if internally F, Z stand for some lists). It's okay to leave F[i] to mean ith function in the summation.